wip: In controller, on StateChanged, notify the model, instead of the view. Other clean; init value.

+++ src/main/java/fr/norsys/nutriscore/controller/NutriScoreSliderChangeListener.java
@@ -1,7 +1,5 @@
 package fr.norsys.nutriscore.controller;
 
-import fr.norsys.nutriscore.view.NutriScoreSlider;
-
 import javax.swing.*;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
@@ -14,7 +12,7 @@ public class NutriScoreSliderChangeListener implements ChangeListener {
     public void stateChanged(ChangeEvent e) {
         JSlider source = (JSlider)e.getSource();
         if (!source.getValueIsAdjusting()) {
-            int sliderValue = (int)source.getValue();
+            int sliderValue = source.getValue();
             System.out.println("Slider value: " + sliderValue);
             //TODO controller updateImage(sliderValue);
         }
